backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
pymongo==4.6.3
mongomock==4.1.2
setuptools==69.5.1
bcrypt==4.1.2
python-dotenv==1.0.1
twilio==9.0.4
